# docs/swagger.yaml
basePath: /api
info:
  contact: {}
  description: API for managing groceries
  title: Grocery API
  version: "1.0"
paths:
...
  /DeleteGrocery:
    delete:
      description: Delete a grocery item by providing its ID
      operationId: delete-grocery
      parameters:
      - description: ID of the grocery item to delete
        in: query
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a grocery item
  /UpdateGrocery:
    patch:
      consumes:
      - application/json
      description: Update a grocery item by providing its ID and new data
      operationId: update-grocery
      parameters:
      - description: ID of the grocery item to update
        in: query
        name: id
        required: true
        type: string
      - description: JSON data containing updated information
        in: formData
        name: json-data
        required: true
        type: string
      - description: Image file for the grocery item
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID or missing JSON data'
          schema:
            type: string
        "404":
          description: 'Not Found: Grocery item not found'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update a grocery item
  /ViewAllGroceries:
    get:
      consumes:
      - application/json
      description: Retrieve a list of groceries with optional filters and pagination.
      operationId: view-all-groceries
      parameters:
      - description: Page token for cursor-based pagination
        in: query
        name: pageToken
        type: string
      - description: Filter by product name
        in: query
        name: productname
        type: string
      - description: 'Price filter format: ''gt:100'', ''eq:50'', ''lt:200'''
        in: query
        name: priceFilter
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: View all groceries
...
swagger: "2.0"